refactor(components): migrate 3 more SFCs to h() render functions
배치 2 전환 대상:
- KrdsResize (159줄)
- KrdsContextualHelp (160줄)
- KrdsLanguageSwitcher (186줄)

세 컴포넌트 모두 템플릿 ref 와 전역 이벤트 리스너를 쓰고 있어, ref 를
h() props 로 넘기고 setup 스코프에 그대로 유지했다. @click.prevent 는
withModifiers 로 옮겼다.

남은 SFC 3개: KrdsTooltip, KrdsPagination, KrdsDateInput
